damage-stack: use new standard system
