+++ Blasphemous.LostDreams/Effects/IMultiplierEffect.cs
@@ -1,0 +1,7 @@
+﻿
+namespace Blasphemous.LostDreams.Effects;
+
+internal interface IMultiplierEffect
+{
+    public float Multiplier { get; }
+}
